MeshExportInfo.GetInfo
