--- Assets/UniGLTF/Editor/UniGLTF/ExportDialog/ExporterExtensions.cs
@@ -1,18 +1,0 @@
-using System.Linq;
-using UnityEngine;
-
-namespace UniGLTF
-{
-    public static class ExporterExtensions
-    {
-        public static bool EnableForExport(this Component mono)
-        {
-            if (mono.transform.Ancestors().Any(x => !x.gameObject.activeSelf))
-            {
-                // 自分か祖先に !activeSelf がいる
-                return false;
-            }
-            return true;
-        }
-    }
-}
