--- Assets/UniGLTF/Editor/UniGLTF/ExportDialog/ExporterExtensions.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f76ebfda1e249cb48890de41cc221889
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
